Introduce wasm32-unknown-unknown compilation support
Investigate and resolve architecture-specific compilation barriers preventing
Expander crates from targeting wasm32-unknown-unknown. Changes span field
arithmetic, proving infrastructure, and serialization layers.

- Gate rayon behind optional `parallel` feature in sumcheck, gkr, poly_commit
- Gate non-BN254 field crates behind optional `all_fields` feature in gkr_engine, gkr
- Add scalar fallback implementations for SIMD types (M31x16, BabyBearx16,
  GF2x128, GF2_128, GF2_128x8) on non-x86_64/non-aarch64 architectures
- Broaden Goldilocks NEON (pure scalar) implementation as generic non-x86 fallback
- Gate inline assembly in goldilocks for x86_64/aarch64 only
- Add PackedMontyParameters fallback trait for non-x86_64/non-aarch64
- Replace usize VERSION_NUM constants with u64 for wasm32 compatibility
- Use u64 intermediate in transcript usize vector generation
- Serialize usize as u64 in ExpSerde for cross-platform wire format
- Disable criterion default features to avoid rayon on wasm32
- Add wasm32 NAME constant fallbacks for BabyBearExt3x16, M31Ext3x16

+++ circuit/src/serde.rs
@@ -104,7 +104,7 @@ impl<C: FieldEngine> ExpSerde for Segment<C> {
     }
 }
 
-const VERSION_NUM: usize = 3914834606642317635; // b'CIRCUIT6'
+const VERSION_NUM: u64 = 3914834606642317635; // b'CIRCUIT6'
 
 impl<C: FieldEngine> ExpSerde for RecursiveCircuit<C> {
     fn serialize_into<W: Write>(&self, mut writer: W) -> SerdeResult<()> {
@@ -122,7 +122,7 @@ impl<C: FieldEngine> ExpSerde for RecursiveCircuit<C> {
     }
 
     fn deserialize_from<R: Read>(mut reader: R) -> SerdeResult<Self> {
-        let version_num = <usize as ExpSerde>::deserialize_from(&mut reader).unwrap();
+        let version_num = <u64 as ExpSerde>::deserialize_from(&mut reader).unwrap();
         assert_eq!(version_num, VERSION_NUM);
         let expected_mod = <C::CircuitField as Field>::MODULUS;
         let mut read_mod = [0u8; 32];
